Readjustment to Classes & Moved to matching Folder
