+++ src/classes/misc/counter.ts
